requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
sqlalchemy>=2.0.0
pandas>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
great-expectations>=0.18.0
colorlog>=6.7.0
langchain-ollama>=0.1.0
litellm>=1.0.0
